requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
djangorestframework==3.11.1
gunicorn==20.0.4
markdown2==2.3.9
peewee==3.13.3
protobuf==3.12.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
